+++ SpreadCheetah.SourceGenerator/Models/ColumnWidth.cs
@@ -1,3 +1,3 @@
 namespace SpreadCheetah.SourceGenerator.Models;
 
-internal readonly record struct ColumnWidth(int Width);
+internal readonly record struct ColumnWidth(double Width);
