+++ SpreadCheetah/SourceGeneration/ColumnWidthAttribute.cs
@@ -1,4 +1,4 @@
 namespace SpreadCheetah.SourceGeneration;
 
 [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-public sealed class ColumnWidthAttribute(int width) : Attribute;
+public sealed class ColumnWidthAttribute(double width) : Attribute;
